% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{align_models}
\alias{align_models}
\title{Title Align models.}
\usage{
align_models(..., shared_vocab_only = TRUE)
}
\arguments{
\item{...}{A number of VectorSpaceModels to align as independent entries;
or, alternatively, a single list
of several vectorspace models to align.}

\item{shared_vocab_only}{Two models may have different words. If this variable is TRUE,
(the default), the *entire* vocabular will be returned.}
}
\value{
A list of VectorSpaceModel objects. Each is rotated into the space defined by the first model,
and altered so that each vector is of unit length.
}
\description{
Title Align models.
}
