% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixFunctions.R
\name{as.VectorSpaceModel}
\alias{as.VectorSpaceModel}
\title{Convert to a Vector Space Model}
\usage{
as.VectorSpaceModel(matrix)
}
\arguments{
\item{matrix}{A matrix to coerce.}
}
\value{
An object of class "VectorSpaceModel"
}
\description{
Convert to a Vector Space Model
}

