% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixFunctions.R
\name{cosineDist}
\alias{cosineDist}
\title{Cosine Distance}
\usage{
cosineDist(x, y)
}
\arguments{
\item{x}{A matrix, VectorSpaceModel, or vector.}

\item{y}{A matrix, VectorSpaceModel, or vector.}
}
\value{
A matrix whose dimnames are rownames(x), rownames(y) and whose entires are
the associated distance.
}
\description{
Calculate the cosine distance between two vectors.

Not an actual distance metric, but can be used in similar contexts.
It is calculated as simply the inverse of cosine similarity,
and falls in a fixed range of 0 (identical) to 2 (completely opposite in direction.)
}
