% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrixFunctions.R
\name{cosineSimilarity}
\alias{cosineSimilarity}
\title{Cosine Similarity}
\usage{
cosineSimilarity(x, y)
}
\arguments{
\item{x}{A matrix or VectorSpaceModel object}

\item{y}{A vector, matrix or VectorSpaceModel object.
Vectors are coerced to single-row matrices; y must have the
same number of dimensions as x.}
}
\value{
A matrix with one row for each row of x. Columns and their names correspond to y.
}
\description{
Calculate the cosine similarity of two matrices or a matrix and a vector.
}

