% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{extract_underlying_vector}
\alias{extract_underlying_vector}
\title{Extract a shared vector from a list of vectors.}
\usage{
extract_underlying_vector(matrix, ...)
}
\arguments{
\item{...}{A list of vectors, words, or formulas to extract in the context of "matrix."}
}
\value{
A single vector representing the principal component of the various input vector.
}
\description{
Extract a shared vector from a list of vectors.
}
\examples{
gender_vector_from_pca <- demo_vectors \%>\% extract_underlying_vector(~"he"-"she",~"man"-"woman",~"himself"-"herself",~("mrs"+"ms")/2 - "mr")
demo_vectors \%>\% nearest_to(gender_vector_from_pca)

}
