% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixFunctions.R
\name{extract_vector_from_pairs}
\alias{extract_vector_from_pairs}
\title{Extract a single vector from a list of pairs.}
\usage{
extract_vector_from_pairs(matrix, pairs)
}
\arguments{
\item{matrix}{A `VectorSpaceModel` object.}

\item{pairs}{A list of pairs of words contained in the vectorset.
Each pair should be positioned on opposite ends of the vector
For example, use
\code{list(c("he","she"),c("man","woman"),c("him","her"))} to
get a "gender vector" that may be slightly better than just the average
of the three vectors.}
}
\value{
A vector of the same dimensionality as matrix. It will point,
on average, from the first element in each pair to the second.
}
\description{
Extract a single vector from a list of pairs.
}
\examples{

gender = demo_vectors \%>\%
   extract_vector_from_pairs(list(c("he","she"),c("man","woman"),c("him","her")))

demo_vectors \%>\% closest_to(gender)

}
