% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixFunctions.R
\name{extract_vectors}
\alias{extract_vectors}
\title{Extract vectors or formulas from a VectorSpaceModel.}
\usage{
extract_vectors(vsm, ...)
}
\arguments{
\item{vsm}{A VectorSpaceModel or matrix}

\item{...}{Words or formulas to extract from. Alternatively, a single list
of formulas.}
}
\value{
A Vector space model
}
\description{
Extract vectors or formulas from a VectorSpaceModel.
}
\examples{

demo_vectors \%>\% 
  extract_vectors("lady","woman","man","he","she","guy","man")
  
}
