% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrixFunctions.R
\name{filter_to_rownames}
\alias{filter_to_rownames}
\title{Reduce by rownames}
\usage{
filter_to_rownames(matrix, words)
}
\arguments{
\item{matrix}{A matrix or VectorSpaceModel object}

\item{words}{A list of rownames or VectorSpaceModel names}
}
\value{
An object of the same class as matrix, consisting
of the rwos that match its rownames.
}
\description{
Reduce by rownames
}

