% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixFunctions.R
\name{improve_vectorspace}
\alias{improve_vectorspace}
\title{Improve Vectorspace}
\usage{
improve_vectorspace(vectorspace, D = round(ncol(vectorspace)/100))
}
\arguments{
\item{vectorspace}{A VectorSpacemodel to be improved.}

\item{D}{The number of principal components to eliminate.}
}
\value{
A VectorSpaceModel object, transformed from the original.
}
\description{
Improve Vectorspace
}
