% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrixFunctions.R
\name{magnitudes}
\alias{magnitudes}
\title{Vector Magnitudes}
\usage{
magnitudes(matrix)
}
\arguments{
\item{matrix}{A matrix or VectorSpaceModel object.}
}
\value{
A vector consisting of the magnitudes of each row.

This is an extraordinarily simple function.
}
\description{
Vector Magnitudes
}

