% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixFunctions.R
\name{nearest_to}
\alias{nearest_to}
\title{Nearest vectors to a word}
\usage{
nearest_to(...)
}
\arguments{
\item{...}{See `closest_to`}
}
\value{
a names vector of cosine similarities. See 'nearest_to' for more details.
}
\description{
This a wrapper around closest_to, included for back-compatibility. Use
closest_to for new applications.
}
\examples{

# Recommended usage in 1.0:
nearest_to(demo_vectors, demo_vectors[["good"]])

# Recommended usage in 2.0:
demo_vectors \%>\% closest_to("good")

}
