% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixFunctions.R
\name{normalize_lengths}
\alias{normalize_lengths}
\title{Matrix normalization.}
\usage{
normalize_lengths(matrix)
}
\arguments{
\item{matrix}{A matrix or VectorSpaceModel object}
}
\value{
An object of the same class as matrix
}
\description{
Normalize a matrix so that all rows are of unit length.
}

