% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/word2vec.R
\name{prep_word2vec}
\alias{prep_word2vec}
\title{Prepare documents for word2Vec}
\usage{
prep_word2vec(origin, destination, split_characters = "\\\\W",
  lowercase = F)
}
\arguments{
\item{origin}{A text file or a directory of text files
to be used in training the model}

\item{destination}{The location for output text.}

\item{split_characters}{A list of characters that mark word breaks. By default,
any nonword characters according to the perl regex engine.}

\item{tolower}{Logical. Should uppercase characters be converted to lower?}
}
\value{
The file name (silently).
}
\description{
This function exports a directory or document to a single file
suitable to Word2Vec run on. That means a single, seekable txt file
with tokens separated by spaces. (For example, punctuation is removed
rather than attached to the end of words.)
This function is quite inefficient: in most cases, you'll be better
off preparing the documents using.
}

