% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{procrustes}
\alias{procrustes}
\title{Align two matrices with orthogonal procrustes}
\usage{
procrustes(x, y)
}
\arguments{
\item{x}{A reference matrix or VectorSpaceModel.}

\item{y}{The matrix or VectorSpaceModel to be rotated.}
}
\value{
A matrix of the same shape as y, projected into the space of x.
}
\description{
Orthogonal procrustes determines a matrix rotation that
best positions points from one arbitrary vectorspace into another. This
function uses rownames to find which rows should be aligned with each other.
Note that this may produce undesired results if the rownames are not truly
the same words; for example, if the models were changed with different tokenization
schemes or rules for ngram bundling.

William L. Hamilton, Jure Leskovec, and Dan Jurafsky.
2016. Diachronic Word Embeddings Reveal Statistical Laws of Semantic Change.
Proceedings of ACL.
}
\examples{

# Create a new model from the reference with gender rejected out.
genderless <- demo_vectors \%>\% reject(~"man" - "woman") \%>\% reject(~"he" - "she")

# The vector for "she" in the new space is only somewhat similar to the old one.
demo_vectors[["she"]] \%>\% cosineSimilarity(genderless[["she"]])

# We can re-align the genderless vectorspace into the original one.
adjusted <- demo_vectors \%>\% procrustes(genderless)

# Now the cosine similarity is ever-so-slightly higher.
demo_vectors[["she"]] \%>\% cosineSimilarity(adjusted[["she"]])
}
