% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixFunctions.R
\name{project}
\alias{project}
\title{Project each row of an input matrix along a vector.}
\usage{
project(matrix, vector)
}
\arguments{
\item{matrix}{A matrix or VectorSpaceModel}

\item{vector}{A vector (or object coercable to a vector)
of the same length as the VectorSpaceModel.}
}
\value{
A new matrix or VectorSpaceModel of the same dimensions as `matrix`,
each row of which is parallel to vector.

If the input is a matrix, the output will be a matrix: if a VectorSpaceModel,
it will be a VectorSpaceModel.
}
\description{
As with 'cosineSimilarity
}

