% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{read_group}
\alias{read_group}
\title{Read Group}
\usage{
read_group(filenames, ...)
}
\arguments{
\item{filenames}{A list of filenames naming VectorSpaceModels.}

\item{...}{Additional parameters passed to read_vectors}
}
\value{
A lisit of VectorSpaceModel objects
}
\description{
Read Group
}
\examples{

\dontrun{
group = read_group(c("model1.bin", "model2.bin", "model3.bin"))
aligned = align_models(group)
}

}
