% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixFunctions.R
\name{square_magnitudes}
\alias{square_magnitudes}
\title{Square Magnitudes with caching}
\usage{
square_magnitudes(object)
}
\arguments{
\item{VectorSpaceModel}{A matrix or VectorSpaceModel object}
}
\value{
A vector of the square magnitudes for each row
}
\description{
square_magnitudes Returns the square magnitudes and
caches them if necessary
}
\keyword{internal}

