% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{stanford_plot}
\alias{stanford_plot}
\title{Plot models}
\usage{
stanford_plot(model_set, word, n = 20, transform_type = c("mds", "pca"))
}
\arguments{
\item{transform_type}{}
}
\description{
Plot models
}
