% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrixFunctions.R
\docType{methods}
\name{[[,VectorSpaceModel-method}
\alias{[[,VectorSpaceModel-method}
\title{VectorSpaceModel subsetting}
\usage{
\S4method{[[}{VectorSpaceModel}(x, i, average = TRUE, ...)
}
\arguments{
\item{i}{The words to use as rownames}

\item{average}{Whether to collapse down to a single vector,
or to return a subset of one row for each asked for.}

\item{VectorSpaceModel}{}
}
\value{
A VectorSpaceModel of a single row.
}
\description{
VectorSpaceModel subsetting
}

