% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrixFunctions.R
\docType{methods}
\name{[,VectorSpaceModel-method}
\alias{[,VectorSpaceModel-method}
\title{VectorSpaceModel indexing}
\usage{
\S4method{[}{VectorSpaceModel}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{VectorSpaceModel}{}
}
\value{
A VectorSpaceModel

I believe this is necessary, but honestly am not sure.
}
\description{
VectorSpaceModel indexing
}

