% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/word2vec.R
\name{train_word2vec}
\alias{train_word2vec}
\title{Train a model by word2vec.}
\usage{
train_word2vec(train_file, output_file = "vectors.txt", vectors = 100,
  threads = 1, window = 12)
}
\arguments{
\item{train_file}{Path of a single .txt file for training. Tokens are split on spaces.}

\item{output_file}{Path of the output file.}
}
\value{
A word2vec object.
}
\description{
Train a model by word2vec.
}
\details{
The word2vec tool takes a text corpus as input and produces the
word vectors as output. It first constructs a vocabulary from the
training text data and then learns vector representation of words.
The resulting word vector file can be used as features in many
natural language processing and machine learning applications.
}
\examples{
\dontrun{
model = word2vec(system.file("examples", "rfaq.txt", package = "tmcn.word2vec"))
}
}
\author{
Jian Li <\email{rweibo@sina.com}>, Ben Schmidt <\email{bmchmidt@gmail.com}>
}
\references{
\url{https://code.google.com/p/word2vec/}
}

