% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrixFunctions.R
\docType{class}
\name{VectorSpaceModel-class}
\alias{VectorSpaceModel-class}
\title{Vector Space Model class}
\value{
An object of class "VectorSpaceModel"
}
\description{
A class for describing and accessing Vector Space Models like Word2Vec.
The base object is simply a matrix with columns describing dimensions and unique rownames
as the names of vectors. This package gives a number of convenience functions for printing
and, most importantly, accessing these objects.
}

