% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixFunctions.R
\docType{methods}
\name{-,VectorSpaceModel,VectorSpaceModel-method}
\alias{-,VectorSpaceModel,VectorSpaceModel-method}
\title{VectorSpaceModel subtraction}
\usage{
\S4method{-}{VectorSpaceModel,VectorSpaceModel}(e1, e2)
}
\arguments{
\item{e1}{A vector space model}

\item{e2}{A vector space model of equal size OR a vector
space model of a single row. If the latter (which is more likely)
the specified row will be subtracted from each row.}
}
\value{
A VectorSpaceModel of the same dimensions and rownames
as e1

I believe this is necessary, but honestly am not sure.
}
\description{
Keep the VSM class when doing subtraction operations;
make it possible to subtract a single row from an entire model.
}

