% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_function.R
\name{write.txt.word2vec}
\alias{write.txt.word2vec}
\title{Write in word2vec text format}
\usage{
write.txt.word2vec(model, filename)
}
\arguments{
\item{model}{The wordVectors model you wish to save. (This can actually be any matrix with rownames,
if you want a smaller binary serialization in single-precision floats.)}

\item{filename}{The file to save the vectors to. I recommend ".vectors" as a suffix.}
}
\value{
Nothing
}
\description{
Write in word2vec text format
}
