% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrixFunctions.R
\name{cosineSimilarity}
\alias{cosineSimilarity}
\title{Cosine Similarity}
\usage{
cosineSimilarity(x, y)
}
\arguments{
\item{x}{A matrix or VectorSpaceModel object}

\item{y}{A vector, matrix or VectorSpaceModel object.

Vector inputs are coerced to single-row matrices; y must have the
same number of dimensions as x.}
}
\value{
A matrix. Rows correspond to entries in x; columns to entries in y.
}
\description{
Calculate the cosine similarity of two matrices or a matrix and a vector.
}
\examples{
subjects = demo_vectors[[c("history","literature","biology","math","stats"),average=FALSE]]
cosineSimilarity(subjects,subjects)
}

