% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixFunctions.R
\name{extract_vectors}
\alias{extract_vectors}
\title{Extract vectors or formulas from a VectorSpaceModel.}
\usage{
extract_vectors(vsm, ..., labels = NULL)
}
\arguments{
\item{vsm}{A VectorSpaceModel or matrix}

\item{...}{Words or formulas to extract from. Alternatively, a single list
of formulas.}

\item{...}{Optionally, slabels to use for the vector space model: a character vector the same length
as vectors to extract. This may be useful in programming contexts.}
}
\value{
A Vector space model with the same number of rows as \code{...}
}
\description{
Extract vectors or formulas from a VectorSpaceModel.
}
\examples{

demo_vectors \%>\%
  extract_vectors("woman","man", ~"man" + "woman",  ~"man" - "woman")

}
