% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixFunctions.R
\name{improve_vectorspace}
\alias{improve_vectorspace}
\title{Improve a vectorspace by removing common elements.}
\usage{
improve_vectorspace(vectorspace, D = round(ncol(vectorspace)/100))
}
\arguments{
\item{vectorspace}{A VectorSpacemodel to be improved.}

\item{D}{The number of principal components to eliminate.}
}
\value{
A VectorSpaceModel object, transformed from the original.
}
\description{
See reference for a full description. Supposedly, these operations will improve performance on analogy tasks.
}
\examples{
cosin
closest_to(demo_vectors,"great")
# stopwords like "and" and "very" are no longer top ten.
# I don't know if this is really better, though.

closest_to(improve_vectorspace(demo_vectors),"great")

}
\references{
Jiaqi Mu, Suma Bhat, Pramod Viswanath. All-but-the-Top: Simple and Effective Postprocessing for Word Representations. https://arxiv.org/abs/1702.01417.
}
