% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixFunctions.R
\name{nearest_to}
\alias{nearest_to}
\title{Return the n closest words in a VectorSpaceModel to a given vector.}
\usage{
nearest_to(matrix, vector, n = 10)
}
\arguments{
\item{matrix}{A matrix or VectorSpaceModel}

\item{vector}{A vector (or an object coercable to a vector, see project)
of the same length as the VectorSpaceModel. Or, for convenience a string
representing a word in the passed matrix. See examples.}

\item{n}{The number of closest words to include.}
}
\value{
A vector of distances, with names corresponding to the words
in the parent VectorSpaceModel, of length n.
}
\description{
Return the n closest words in a VectorSpaceModel to a given vector.
}
\examples{

# Synonyms and similar words
nearest_to(demo_vectors,demo_vectors[["good"]])

If 'matrix' is a VectorSpaceModel object,
you can also just enter a string directly, and
So the results below hold.

nearest_to(demo_vectors,"good")

# Something close to the classic king:man::queen:woman;
# What's the equivalent word for a female teacher that "guy" is for
# a male one?

nearest_to(demo_vectors,"guy" - "man" + "woman")

}

