% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixFunctions.R
\docType{methods}
\name{plot,VectorSpaceModel-method}
\alias{plot,VectorSpaceModel-method}
\title{Plot a Vector Space Model.}
\usage{
\S4method{plot}{VectorSpaceModel}(x, y, ...)
}
\arguments{
\item{x}{The model to plot}

\item{y}{(ignored)}

\item{...}{Further arguments passed to tsne::tsne.
(Note: not to plot.)}
}
\value{
The TSNE model (silently.)
}
\description{
Visualizing a model as a whole is sort of undefined. I think the
sanest thing to do is reduce the full model down to two dimensions
using T-SNE, which preserves some of the local clusters.
}
\details{
This plots only the first 300 words in the model.
}

