% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixFunctions.R
\docType{methods}
\name{plot,VectorSpaceModel-method}
\alias{plot,VectorSpaceModel-method}
\title{Plot a Vector Space Model.}
\usage{
\S4method{plot}{VectorSpaceModel}(x, method = "tsne", ...)
}
\arguments{
\item{x}{The model to plot}

\item{method}{The method to use for plotting. "pca" is principal components, "tsne" is t-sne}

\item{...}{Further arguments passed to the plotting method.}
}
\value{
The TSNE model (silently.)
}
\description{
Visualizing a model as a whole is sort of undefined. I think the
sanest thing to do is reduce the full model down to two dimensions
using T-SNE, which preserves some of the local clusters.
}
\details{
For individual subsections, it can make sense to do a principal components
plot of the space of just those letters. This is what happens if method
is pca. On the full vocab, it's kind of a mess.

This plots only the first 300 words in the model.
}

