% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/word2vec.R
\name{prep_word2vec}
\alias{prep_word2vec}
\title{Prepare documents for word2Vec}
\usage{
prep_word2vec(origin, destination, split_characters = "\\\\W",
  lowercase = F)
}
\arguments{
\item{origin}{A text file or a directory of text files
to be used in training the model}

\item{destination}{The location for output text.}

\item{split_characters}{If the 'stringi' package is not installed,
A list of characters that mark word breaks. By default,
any nonword characters according to the perl regex engine. If stringi is installed,
this parameter is ignored.}

\item{tolower}{Logical. Should uppercase characters be converted to lower?}
}
\value{
The file name (silently).
}
\description{
This function exports a directory or document to a single file
suitable to Word2Vec run on. That means a single, seekable txt file
with tokens separated by spaces. (For example, punctuation is removed
rather than attached to the end of words.)
This function is extraordinarily inefficient: in most real-world cases, you'll be
much better off preparing the documents using python, perl, awk, or any other
scripting language that can reasonable read things in line-by-line.
}

