% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrixFunctions.R
\name{read.binary.vectors}
\alias{read.binary.vectors}
\title{Read binary word2vec format files}
\usage{
read.binary.vectors(filename, nrows = Inf)
}
\arguments{
\item{filename}{A file in the binary word2vec format to import.}

\item{nrows}{Optionally, a number of rows to stop reading after.
Word2vec sorts by frequency, so limiting to the first 1000 rows will
give the thousand most-common words; it can be useful not to load
the whole matrix into memory}
}
\value{
A word2vec object
}
\description{
Read binary word2vec format files
}

