% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixFunctions.R
\name{read.vectors}
\alias{read.vectors}
\title{Read VectorSpaceModel}
\usage{
read.vectors(filename, vectors = NULL, binary = NULL, header_row = TRUE,
  ...)
}
\arguments{
\item{filename}{The file to read in.}

\item{vectors}{The number of dimensions word2vec calculated. Imputed automatically if not specified.}

\item{binary}{Read in the binary word2vec form. (Wraps `read.binary.vectors`) By default, function
guesses based on file suffix.}

\item{header_row}{Whether the first row of the file gives information about the rest. If FALSE,
the number of dimensions is inferred from the first row. Useful in reading gloVe files.}

\item{...}{Further arguments passed to read.table or read.binary.vectors.
Note that both accept 'nrows' as an argument. Word2vec produces
by default frequency sorted output. Therefore 'read.vectors("file.bin", nrows=500)', for example,
will return the vectors for the top 500 words. This can be useful on machines with limited
memory.}
}
\value{
An matrixlike object of class `VectorSpaceModel`
}
\description{
Read a VectorSpaceModel from a file exported from word2vec or a similar output format.
}
