% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrixFunctions.R
\name{read.vectors}
\alias{read.vectors}
\title{Read VectorSpaceModel}
\usage{
read.vectors(filename, vectors = guess_n_cols(), ...)
}
\arguments{
\item{filename}{The file to read in.}

\item{vectors}{The number of dimensions word2vec calculated. Imputed automatically if not specified.}

\item{...}{Further arguments passed to read.table. Word2vec produces
by default frequency sorted output. Therefore [limits nrow=500], for example,
will return the vectors to the top 500 words.}
}
\value{
An matrixlike object of class `VectorSpaceModel`
}
\description{
Read a VectorSpaceModel from a file exported from word2vec or a similar output format.
}

