% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixFunctions.R
\name{reject}
\alias{reject}
\title{Return a vector rejection for each element in a VectorSpaceModel}
\usage{
reject(matrix, vector)
}
\arguments{
\item{matrix}{A matrix or VectorSpaceModel}

\item{vector}{A vector (or an object coercable to a vector, see project)
of the same length as the VectorSpaceModel.}
}
\value{
A new matrix or VectorSpaceModel of the same dimensions as `matrix`,
each row of which is orthogonal to the `vector` object.

This is defined simply as `matrix-project(matrix,vector)`, but having a separate
name may make for cleaner code.

See `project` for more details.
}
\description{
Return a vector rejection for each element in a VectorSpaceModel
}
\examples{
nearest_to(demo_vectors,demo_vectors[["man"]])
genderless = reject(demo_vectors,demo_vectors[["he"]] - demo_vectors[["she"]])
nearest_to(genderless,genderless[["man"]])

}

