% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrixFunctions.R
\docType{methods}
\name{[[,VectorSpaceModel-method}
\alias{[[,VectorSpaceModel-method}
\title{VectorSpaceModel subsetting}
\usage{
\S4method{[[}{VectorSpaceModel}(x, i, j, ...)
}
\arguments{
\item{VectorSpaceModel}{}
}
\value{
A VectorSpaceModel of a single row.
}
\description{
VectorSpaceModel subsetting
}

