% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrixFunctions.R
\docType{methods}
\name{[,VectorSpaceModel-method}
\alias{[,VectorSpaceModel-method}
\title{VectorSpaceModel indexing}
\usage{
\S4method{[}{VectorSpaceModel}(x, i, j)
}
\arguments{
\item{x}{The vectorspace model to subset}

\item{i}{The row numbers to extract}

\item{j}{The column numbers to extract}
}
\value{
A VectorSpaceModel
}
\description{
VectorSpaceModel indexing
}

