% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixFunctions.R
\docType{methods}
\name{[,VectorSpaceModel-method}
\alias{[,VectorSpaceModel-method}
\title{VectorSpaceModel indexing}
\usage{
\S4method{[}{VectorSpaceModel}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{The vectorspace model to subset}

\item{i}{The row numbers to extract}

\item{j}{The column numbers to extract}

\item{...}{Other arguments passed to extract (unlikely to be useful).}

\item{drop}{Whether to drop columns. This parameter is ignored.}
}
\value{
A VectorSpaceModel
}
\description{
Reduce a VectorSpaceModel to a smaller one
}

