% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixFunctions.R
\name{write.txt.word2vec}
\alias{write.txt.word2vec}
\title{Write in word2vec text format}
\usage{
write.txt.word2vec(model, filename)
}
\arguments{
\item{model}{The wordVectors model you wish to save.}

\item{filename}{The file to save the vectors to. I recommend ".vectors" or ".txt" as a suffix.}
}
\value{
Nothing
}
\description{
Write in word2vec text format
}
